+++ server/pipeline.go
@@ -20,7 +20,7 @@ import (
 
 	"github.com/golang/protobuf/jsonpb"
 	"github.com/heroiclabs/nakama/rtapi"
-	"github.com/satori/go.uuid"
+	"go.opencensus.io/trace"
 	"go.uber.org/zap"
 )
 
@@ -67,27 +67,77 @@ func (p *Pipeline) ProcessRequest(logger *zap.Logger, session Session, envelope 
 		return false
 	}
 
-	activateHooks := true
-	messageName := ""
-	uid := uuid.Nil
-	username := ""
-	expiry := int64(0)
-	sessionID := ""
+	var pipelineFn func(*zap.Logger, Session, *rtapi.Envelope)
+	var pipelineName string
+
+	switch envelope.Message.(type) {
+	case *rtapi.Envelope_ChannelJoin:
+		pipelineFn = p.channelJoin
+		pipelineName = "channelJoin"
+	case *rtapi.Envelope_ChannelLeave:
+		pipelineFn = p.channelLeave
+		pipelineName = "channelLeave"
+	case *rtapi.Envelope_ChannelMessageSend:
+		pipelineFn = p.channelMessageSend
+		pipelineName = "channelMessageSend"
+	case *rtapi.Envelope_ChannelMessageUpdate:
+		pipelineFn = p.channelMessageUpdate
+		pipelineName = "channelMessageUpdate"
+	case *rtapi.Envelope_ChannelMessageRemove:
+		pipelineFn = p.channelMessageRemove
+		pipelineName = "channelMessageRemove"
+	case *rtapi.Envelope_MatchCreate:
+		pipelineFn = p.matchCreate
+		pipelineName = "matchCreate"
+	case *rtapi.Envelope_MatchDataSend:
+		pipelineFn = p.matchDataSend
+		pipelineName = "matchDataSend"
+	case *rtapi.Envelope_MatchJoin:
+		pipelineFn = p.matchJoin
+		pipelineName = "matchJoin"
+	case *rtapi.Envelope_MatchLeave:
+		pipelineFn = p.matchLeave
+		pipelineName = "matchLeave"
+	case *rtapi.Envelope_MatchmakerAdd:
+		pipelineFn = p.matchmakerAdd
+		pipelineName = "matchmakerAdd"
+	case *rtapi.Envelope_MatchmakerRemove:
+		pipelineFn = p.matchmakerRemove
+		pipelineName = "matchmakerRemove"
+	case *rtapi.Envelope_Rpc:
+		pipelineFn = p.rpc
+		pipelineName = fmt.Sprintf("rpc.%v", envelope.GetRpc().Id)
+	case *rtapi.Envelope_StatusFollow:
+		pipelineFn = p.statusFollow
+		pipelineName = "statusFollow"
+	case *rtapi.Envelope_StatusUnfollow:
+		pipelineFn = p.statusUnfollow
+		pipelineName = "statusUnfollow"
+	case *rtapi.Envelope_StatusUpdate:
+		pipelineFn = p.statusUpdate
+		pipelineName = "statusUpdate"
+	default:
+		// If we reached this point the envelope was valid but the contents are missing or unknown.
+		// Usually caused by a version mismatch, and should cause the session making this pipeline request to close.
+		logger.Error("Unrecognizable payload received.", zap.Any("payload", envelope))
+		session.Send(false, 0, &rtapi.Envelope{Cid: envelope.Cid, Message: &rtapi.Envelope_Error{Error: &rtapi.Error{
+			Code:    int32(rtapi.Error_UNRECOGNIZED_PAYLOAD),
+			Message: "Unrecognized message.",
+		}}})
+		return false
+	}
+
+	var messageName string
 
 	switch envelope.Message.(type) {
 	case *rtapi.Envelope_Rpc:
-		activateHooks = false
+		// No before/after hooks on RPC.
 	default:
 		messageName = fmt.Sprintf("%T", envelope.Message)
-		uid = session.UserID()
-		username = session.Username()
-		expiry = session.Expiry()
-		sessionID = session.ID().String()
-	}
 
-	if activateHooks {
-
-		hookResult, hookErr := invokeReqBeforeHook(logger, p.config, p.runtimePool, p.jsonpbMarshaler, p.jsonpbUnmarshaler, sessionID, uid, username, expiry, messageName, envelope)
+		span := trace.NewSpan(fmt.Sprintf("nakama.rtapi-before.%v", pipelineName), nil, trace.StartOptions{})
+		hookResult, hookErr := invokeReqBeforeHook(logger, p.config, p.runtimePool, p.jsonpbMarshaler, p.jsonpbUnmarshaler, session.ID().String(), session.UserID(), session.Username(), session.Expiry(), messageName, envelope)
+		span.End()
 
 		if hookErr != nil {
 			session.Send(false, 0, &rtapi.Envelope{Cid: envelope.Cid, Message: &rtapi.Envelope_Error{Error: &rtapi.Error{
@@ -117,50 +167,14 @@ func (p *Pipeline) ProcessRequest(logger *zap.Logger, session Session, envelope 
 		envelope = resultCast
 	}
 
-	switch envelope.Message.(type) {
-	case *rtapi.Envelope_ChannelJoin:
-		p.channelJoin(logger, session, envelope)
-	case *rtapi.Envelope_ChannelLeave:
-		p.channelLeave(logger, session, envelope)
-	case *rtapi.Envelope_ChannelMessageSend:
-		p.channelMessageSend(logger, session, envelope)
-	case *rtapi.Envelope_ChannelMessageUpdate:
-		p.channelMessageUpdate(logger, session, envelope)
-	case *rtapi.Envelope_ChannelMessageRemove:
-		p.channelMessageRemove(logger, session, envelope)
-	case *rtapi.Envelope_MatchCreate:
-		p.matchCreate(logger, session, envelope)
-	case *rtapi.Envelope_MatchDataSend:
-		p.matchDataSend(logger, session, envelope)
-	case *rtapi.Envelope_MatchJoin:
-		p.matchJoin(logger, session, envelope)
-	case *rtapi.Envelope_MatchLeave:
-		p.matchLeave(logger, session, envelope)
-	case *rtapi.Envelope_MatchmakerAdd:
-		p.matchmakerAdd(logger, session, envelope)
-	case *rtapi.Envelope_MatchmakerRemove:
-		p.matchmakerRemove(logger, session, envelope)
-	case *rtapi.Envelope_Rpc:
-		p.rpc(logger, session, envelope)
-	case *rtapi.Envelope_StatusFollow:
-		p.statusFollow(logger, session, envelope)
-	case *rtapi.Envelope_StatusUnfollow:
-		p.statusUnfollow(logger, session, envelope)
-	case *rtapi.Envelope_StatusUpdate:
-		p.statusUpdate(logger, session, envelope)
-	default:
-		// If we reached this point the envelope was valid but the contents are missing or unknown.
-		// Usually caused by a version mismatch, and should cause the session making this pipeline request to close.
-		logger.Error("Unrecognizable payload received.", zap.Any("payload", envelope))
-		session.Send(false, 0, &rtapi.Envelope{Cid: envelope.Cid, Message: &rtapi.Envelope_Error{Error: &rtapi.Error{
-			Code:    int32(rtapi.Error_UNRECOGNIZED_PAYLOAD),
-			Message: "Unrecognized message.",
-		}}})
-		return false
-	}
+	span := trace.NewSpan(fmt.Sprintf("nakama.rtapi.%v", pipelineName), nil, trace.StartOptions{})
+	pipelineFn(logger, session, envelope)
+	span.End()
 
-	if activateHooks {
-		invokeReqAfterHook(logger, p.config, p.runtimePool, p.jsonpbMarshaler, sessionID, uid, username, expiry, messageName, envelope)
+	if messageName != "" {
+		span := trace.NewSpan(fmt.Sprintf("nakama.rtapi-after.%v", pipelineName), nil, trace.StartOptions{})
+		invokeReqAfterHook(logger, p.config, p.runtimePool, p.jsonpbMarshaler, session.ID().String(), session.UserID(), session.Username(), session.Expiry(), messageName, envelope)
+		span.End()
 	}
 
 	return true