+++ server/runtime.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/heroiclabs/nakama/social"
 	"github.com/yuin/gopher-lua"
+	"go.opencensus.io/stats"
 	"go.uber.org/zap"
 	"google.golang.org/grpc/codes"
 )
@@ -54,9 +55,17 @@ type RuntimePool struct {
 	maxCount     int
 	currentCount int
 	newFn        func() *Runtime
+
+	statsCtx          context.Context
+	statsRuntimeCount *stats.Int64Measure
 }
 
 func NewRuntimePool(logger, startupLogger *zap.Logger, db *sql.DB, config Config, socialClient *social.Client, leaderboardCache LeaderboardCache, sessionRegistry *SessionRegistry, matchRegistry MatchRegistry, tracker Tracker, router MessageRouter, stdLibs map[string]lua.LGFunction, moduleCache *ModuleCache, regCallbacks *RegCallbacks, once *sync.Once) *RuntimePool {
+	statsRuntimeCount, err := stats.Int64("nakama.runtime.count", "Number of pooled runtime instances.", stats.UnitNone)
+	if err != nil {
+		startupLogger.Fatal("Error creating stats entry for runtime count", zap.Error(err))
+	}
+
 	rp := &RuntimePool{
 		logger:       logger,
 		regCallbacks: regCallbacks,
@@ -72,6 +81,8 @@ func NewRuntimePool(logger, startupLogger *zap.Logger, db *sql.DB, config Config
 			}
 			return r
 		},
+		statsCtx:          context.Background(),
+		statsRuntimeCount: statsRuntimeCount,
 	}
 
 	// Warm up the pool.
@@ -81,6 +92,7 @@ func NewRuntimePool(logger, startupLogger *zap.Logger, db *sql.DB, config Config
 		for i := 0; i < config.GetRuntime().MinCount; i++ {
 			rp.poolCh <- rp.newFn()
 		}
+		stats.Record(rp.statsCtx, rp.statsRuntimeCount.M(int64(config.GetRuntime().MinCount)))
 	}
 	startupLogger.Info("Allocated minimum runtime pool")
 
@@ -116,10 +128,19 @@ func (rp *RuntimePool) Get() *Runtime {
 			// If we've reached the max allowed allocation block on an available runtime.
 			return <-rp.poolCh
 		}
-		// Allocate a new runtime.
-		rp.currentCount++
-		rp.Unlock()
-		return rp.newFn()
+		// Inside the locked region now, last chance to use an available idle runtime.
+		// Note: useful in case a runtime becomes available while waiting to acquire lock.
+		select {
+		case r := <-rp.poolCh:
+			rp.Unlock()
+			return r
+		default:
+			// Allocate a new runtime.
+			rp.currentCount++
+			rp.Unlock()
+			stats.Record(rp.statsCtx, rp.statsRuntimeCount.M(1))
+			return rp.newFn()
+		}
 	}
 }
 
@@ -137,8 +158,8 @@ func (rp *RuntimePool) Put(r *Runtime) {
 func newVM(logger *zap.Logger, db *sql.DB, config Config, socialClient *social.Client, leaderboardCache LeaderboardCache, sessionRegistry *SessionRegistry, matchRegistry MatchRegistry, tracker Tracker, router MessageRouter, stdLibs map[string]lua.LGFunction, moduleCache *ModuleCache, once *sync.Once, announceCallback func(ExecutionMode, string)) (*Runtime, error) {
 	// Initialize a one-off runtime to ensure startup code runs and modules are valid.
 	vm := lua.NewState(lua.Options{
-		CallStackSize:       1024,
-		RegistrySize:        1024,
+		CallStackSize:       config.GetRuntime().CallStackSize,
+		RegistrySize:        config.GetRuntime().RegistrySize,
 		SkipOpenLibs:        true,
 		IncludeGoStackTrace: true,
 	})
@@ -237,7 +258,7 @@ func (r *Runtime) loadModules(moduleCache *ModuleCache) error {
 }
 
 func (r *Runtime) NewStateThread() (*lua.LState, context.CancelFunc) {
-	return r.vm.NewThread()
+	return r.vm, nil
 }
 
 func (r *Runtime) GetCallback(e ExecutionMode, key string) *lua.LFunction {
@@ -257,16 +278,13 @@ func (r *Runtime) GetCallback(e ExecutionMode, key string) *lua.LFunction {
 }
 
 func (r *Runtime) InvokeFunction(execMode ExecutionMode, fn *lua.LFunction, queryParams map[string][]string, uid string, username string, sessionExpiry int64, sid string, payload interface{}) (interface{}, error, codes.Code) {
-	l, _ := r.NewStateThread()
-	defer l.Close()
-
-	ctx := NewLuaContext(l, r.luaEnv, execMode, queryParams, uid, username, sessionExpiry, sid)
+	ctx := NewLuaContext(r.vm, r.luaEnv, execMode, queryParams, uid, username, sessionExpiry, sid)
 	var lv lua.LValue
 	if payload != nil {
-		lv = ConvertValue(l, payload)
+		lv = ConvertValue(r.vm, payload)
 	}
 
-	retValue, err, code := r.invokeFunction(l, fn, ctx, lv)
+	retValue, err, code := r.invokeFunction(r.vm, fn, ctx, lv)
 	if err != nil {
 		return nil, err, code
 	}