+++ tests/util.go
@@ -53,8 +53,9 @@ var (
 
 type DummyMessageRouter struct{}
 
-func (d *DummyMessageRouter) SendToPresenceIDs(*zap.Logger, []*server.PresenceID, *rtapi.Envelope) {}
-func (d *DummyMessageRouter) SendToStream(*zap.Logger, server.PresenceStream, *rtapi.Envelope)     {}
+func (d *DummyMessageRouter) SendToPresenceIDs(*zap.Logger, []*server.PresenceID, bool, uint8, *rtapi.Envelope) {
+}
+func (d *DummyMessageRouter) SendToStream(*zap.Logger, server.PresenceStream, *rtapi.Envelope) {}
 
 type DummySession struct {
 	messages []*rtapi.Envelope
@@ -82,11 +83,11 @@ func (d *DummySession) Consume(func(logger *zap.Logger, session server.Session, 
 func (d *DummySession) Format() server.SessionFormat {
 	return server.SessionFormatJson
 }
-func (d *DummySession) Send(envelope *rtapi.Envelope) error {
+func (d *DummySession) Send(isStream bool, mode uint8, envelope *rtapi.Envelope) error {
 	d.messages = append(d.messages, envelope)
 	return nil
 }
-func (d *DummySession) SendBytes(payload []byte) error {
+func (d *DummySession) SendBytes(isStream bool, mode uint8, payload []byte) error {
 	envelope := &rtapi.Envelope{}
 	jsonpbUnmarshaler.Unmarshal(bytes.NewReader(payload), envelope)
 	d.messages = append(d.messages, envelope)